[Node 4] Upgrade NAN to v2 and upgrade socketwatcher
Also update the Travis CI config to run on all versions of Node. I had to install GCC 4.8+ to get it to compile.

Test Plan: Ran a very basic test on OS X listening on `en0` (had to use sudo, which I believe is normal) and was able to pick up packets and see which MAC addr they were from.